Refactor app registry metadata loading
